[snap.auxil.ExamineSnap] moved to module _dev_config
