+++ pyroSAR/snap/__init__.py
@@ -1,2 +1,3 @@
 from .util import geocode
-from .auxil import ExamineSnap, gpt
+from .auxil import gpt
+from pyroSAR._dev_config import ExamineSnap
